# otelgrpc: StreamClientInterceptor ends spans synchronously