Fixes https://github.com/open-telemetry/opentelemetry-go-contrib/issues/1352

Even though we are deprecating the interceptors, I think we should still have them as https://pkg.go.dev/google.golang.org/grpc/stats is experimental API.