@dashpole 2023-11-09
> Fixes https://github.com/open-telemetry/opentelemetry-go-contrib/pull/4535

Wrong link?  Thats a PR.