@pellared 2023-11-09
> This makes sense to me, but do we know why it was asynchronous in the first place?

I have no idea 😢