@MadVikingGod 2023-11-09
Can we do a benchmark comparison between the two?